[优化] go-zero 添加role rpc

+++ zero/model/api.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Api) (out *Api, err error)
 		Update(ctx context.Context, in *Api) (out *Api, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Api, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Api) (rows int64, err error)
@@ -86,7 +86,7 @@ func (s *defaultApiModel) Update(ctx context.Context, in *Api) (out *Api, err er
 }
 
 // 删除Api记录
-func (s *defaultApiModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultApiModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
