+++ zero/model/category.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Category) (out *Category, err error)
 		Update(ctx context.Context, in *Category) (out *Category, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Category, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Category) (rows int64, err error)
@@ -81,7 +81,7 @@ func (s *defaultCategoryModel) Update(ctx context.Context, in *Category) (out *C
 }
 
 // 删除Category记录
-func (s *defaultCategoryModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultCategoryModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
