+++ zero/model/chat_message.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *ChatMessage) (out *ChatMessage, err error)
 		Update(ctx context.Context, in *ChatMessage) (out *ChatMessage, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *ChatMessage, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*ChatMessage) (rows int64, err error)
@@ -88,7 +88,7 @@ func (s *defaultChatMessageModel) Update(ctx context.Context, in *ChatMessage) (
 }
 
 // 删除ChatMessage记录
-func (s *defaultChatMessageModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultChatMessageModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
