+++ zero/model/chat_session.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *ChatSession) (out *ChatSession, err error)
 		Update(ctx context.Context, in *ChatSession) (out *ChatSession, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *ChatSession, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*ChatSession) (rows int64, err error)
@@ -84,7 +84,7 @@ func (s *defaultChatSessionModel) Update(ctx context.Context, in *ChatSession) (
 }
 
 // 删除ChatSession记录
-func (s *defaultChatSessionModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultChatSessionModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
