+++ zero/model/menu.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Menu) (out *Menu, err error)
 		Update(ctx context.Context, in *Menu) (out *Menu, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Menu, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Menu) (rows int64, err error)
@@ -88,7 +88,7 @@ func (s *defaultMenuModel) Update(ctx context.Context, in *Menu) (out *Menu, err
 }
 
 // 删除Menu记录
-func (s *defaultMenuModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultMenuModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
