+++ zero/model/operation_log.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *OperationLog) (out *OperationLog, err error)
 		Update(ctx context.Context, in *OperationLog) (out *OperationLog, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *OperationLog, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*OperationLog) (rows int64, err error)
@@ -93,7 +93,7 @@ func (s *defaultOperationLogModel) Update(ctx context.Context, in *OperationLog)
 }
 
 // 删除OperationLog记录
-func (s *defaultOperationLogModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultOperationLogModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
