+++ zero/model/remark.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Remark) (out *Remark, err error)
 		Update(ctx context.Context, in *Remark) (out *Remark, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Remark, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Remark) (rows int64, err error)
@@ -87,7 +87,7 @@ func (s *defaultRemarkModel) Update(ctx context.Context, in *Remark) (out *Remar
 }
 
 // 删除Remark记录
-func (s *defaultRemarkModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultRemarkModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
