+++ zero/model/role.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Role) (out *Role, err error)
 		Update(ctx context.Context, in *Role) (out *Role, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Role, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Role) (rows int64, err error)
@@ -86,7 +86,7 @@ func (s *defaultRoleModel) Update(ctx context.Context, in *Role) (out *Role, err
 }
 
 // 删除Role记录
-func (s *defaultRoleModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultRoleModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
