+++ zero/model/talk.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Talk) (out *Talk, err error)
 		Update(ctx context.Context, in *Talk) (out *Talk, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Talk, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Talk) (rows int64, err error)
@@ -85,7 +85,7 @@ func (s *defaultTalkModel) Update(ctx context.Context, in *Talk) (out *Talk, err
 }
 
 // 删除Talk记录
-func (s *defaultTalkModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultTalkModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
