+++ zero/model/user_login_history.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *UserLoginHistory) (out *UserLoginHistory, err error)
 		Update(ctx context.Context, in *UserLoginHistory) (out *UserLoginHistory, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *UserLoginHistory, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*UserLoginHistory) (rows int64, err error)
@@ -85,7 +85,7 @@ func (s *defaultUserLoginHistoryModel) Update(ctx context.Context, in *UserLogin
 }
 
 // 删除UserLoginHistory记录
-func (s *defaultUserLoginHistoryModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultUserLoginHistoryModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
