+++ zero/model/user_role.model.go
@@ -17,7 +17,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *UserRole) (out *UserRole, err error)
 		Update(ctx context.Context, in *UserRole) (out *UserRole, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *UserRole, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*UserRole) (rows int64, err error)
@@ -79,7 +79,7 @@ func (s *defaultUserRoleModel) Update(ctx context.Context, in *UserRole) (out *U
 }
 
 // 删除UserRole记录
-func (s *defaultUserRoleModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultUserRoleModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
