+++ zero/model/website_config.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *WebsiteConfig) (out *WebsiteConfig, err error)
 		Update(ctx context.Context, in *WebsiteConfig) (out *WebsiteConfig, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *WebsiteConfig, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*WebsiteConfig) (rows int64, err error)
@@ -82,7 +82,7 @@ func (s *defaultWebsiteConfigModel) Update(ctx context.Context, in *WebsiteConfi
 }
 
 // 删除WebsiteConfig记录
-func (s *defaultWebsiteConfigModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultWebsiteConfigModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
